Moved pl_fit and pl_compare functions to a new file
